requirements.txt
psycopg2
aiobotocore==2.1.2
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
botocore==1.23.24
can-decoder>=0.1.9
canedge-browser>=0.0.8
canmatrix==0.9.5
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
frozenlist==1.3.0
fsspec==2022.2.0
future==0.18.3
idna==3.4
jmespath==0.10.0
mdf-iter>=2.0.5
multidict==6.0.2
numpy==1.24.1
pandas==1.5.3
python-dateutil==2.8.2
pytz==2022.1
s3fs==2022.2.0
six==1.16.0
typing-extensions==4.1.1
urllib3==1.26.9
wrapt==1.14.0
yarl==1.7.2
J1939-PGN==0.4
